Add VbmsDocumentTransactionConcern to DRY things up a bit
